refactor: reuse _subIssues utility function
