+++ library/src/schemas/union/union.ts
@@ -15,9 +15,9 @@ import {
   _addIssue,
   _getStandardProps,
   _joinExpects,
+  _subIssues,
 } from '../../utils/index.ts';
 import type { UnionIssue } from './types.ts';
-import { _subIssues } from './utils/index.ts';
 
 /**
  * Union options type.
