+++ library/src/schemas/union/unionAsync.ts
@@ -16,10 +16,10 @@ import {
   _addIssue,
   _getStandardProps,
   _joinExpects,
+  _subIssues,
 } from '../../utils/index.ts';
 import type { UnionIssue } from './types.ts';
 import type { union } from './union.ts';
-import { _subIssues } from './utils/index.ts';
 
 /**
  * Union options async type.
