+++ library/src/utils/index.ts
@@ -10,6 +10,7 @@ export * from './_isLuhnAlgo/index.ts';
 export * from './_isValidObjectKey/index.ts';
 export * from './_joinExpects/index.ts';
 export * from './_stringify/index.ts';
+export * from './_subIssues/index.ts';
 export * from './entriesFromList/index.ts';
 export * from './entriesFromObjects/index.ts';
 export * from './getDotPath/index.ts';
